+++ NEWS.md
@@ -1,3 +1,5 @@
+# reticulate (development version)
+
 # reticulate 1.40.0
 
 - The S3 classes for some (rarely encountered) Python objects have changed.